New package core added. Core generation logic moved there.
