+++ jdi-dark-rest-generator/src/main/java/com/epam/jdi/generator/v2/core/AbstractJavaCodegenJDI.java
@@ -1,4 +1,4 @@
-package com.epam.jdi.generator.v2;
+package com.epam.jdi.generator.v2.core;
 
 import com.google.common.base.Strings;
 import io.restassured.http.ContentType;
@@ -21,7 +21,7 @@ import org.apache.commons.lang3.StringUtils;
 import java.io.File;
 import java.util.*;
 
-import static com.epam.jdi.generator.v2.JavaCodegenConstantsJDI.JAVA_8_DATE_LIBRARY;
+import static com.epam.jdi.generator.v2.core.JavaCodegenConstantsJDI.JAVA_8_DATE_LIBRARY;
 
 /**
  * Created by oksana_cherniavskaia on 31.08.2020.
