+++ jdi-dark-rest-generator/src/main/java/com/epam/jdi/generator/v2/core/JavaCodegenConstantsJDI.java
@@ -1,4 +1,4 @@
-package com.epam.jdi.generator.v2;
+package com.epam.jdi.generator.v2.core;
 
 import io.swagger.codegen.v3.CodegenConstants;
