--- jdi-dark-rest-generator/src/main/resources/META-INF/services/com.epam.jdi.generator.v2.CodegenConfigJDI
@@ -1,1 +1,0 @@
-com.epam.jdi.generator.v2.JavaClientCodegenJDI
