+++ jdi-dark-rest-generator/src/main/resources/META-INF/services/com.epam.jdi.generator.v2.core.CodegenConfigJDI
@@ -1,0 +1,1 @@
+com.epam.jdi.generator.v2.core.JavaClientCodegenJDI
